Log exceptions to debug.log (#52)
